style: list domain 패키지 구조 변경 (#123)
