+++ src/main/java/com/listywave/image/application/service/ImageService.java
@@ -20,8 +20,8 @@ import com.listywave.image.application.domain.ImageType;
 import com.listywave.image.application.dto.ExtensionRanks;
 import com.listywave.image.application.dto.response.ItemPresignedUrlResponse;
 import com.listywave.image.application.dto.response.UserPresignedUrlResponse;
-import com.listywave.list.application.domain.Item;
-import com.listywave.list.application.domain.ListEntity;
+import com.listywave.list.application.domain.item.Item;
+import com.listywave.list.application.domain.list.ListEntity;
 import com.listywave.list.repository.ItemRepository;
 import com.listywave.list.repository.list.ListRepository;
 import com.listywave.user.application.domain.User;
