+++ src/main/java/com/listywave/list/application/domain/comment/Comment.java
@@ -1,7 +1,8 @@
-package com.listywave.list.application.domain;
+package com.listywave.list.application.domain.comment;
 
 import com.listywave.common.BaseEntity;
-import com.listywave.list.application.vo.Content;
+import com.listywave.list.application.domain.list.Content;
+import com.listywave.list.application.domain.list.ListEntity;
 import com.listywave.user.application.domain.User;
 import jakarta.persistence.Column;
 import jakarta.persistence.Embedded;
