+++ src/main/java/com/listywave/list/application/domain/item/ItemImageUrl.java
@@ -1,4 +1,4 @@
-package com.listywave.list.application.vo;
+package com.listywave.list.application.domain.item;
 
 import com.listywave.common.exception.CustomException;
 import com.listywave.common.exception.ErrorCode;
@@ -15,21 +15,21 @@ import lombok.NoArgsConstructor;
 @Embeddable
 @EqualsAndHashCode
 @NoArgsConstructor(access = AccessLevel.PROTECTED, force = true)
-public class ItemTitle {
+public class ItemImageUrl {
 
-    private static final int LENGTH_LIMIT = 100;
+    private static final int LENGTH_LIMIT = 2048;
 
-    @Column(name = "title", nullable = false, length = LENGTH_LIMIT)
+    @Column(name = "image_url", length = LENGTH_LIMIT)
     private final String value;
 
-    public ItemTitle(String value) {
+    public ItemImageUrl(String value) {
         validate(value);
         this.value = value;
     }
 
     private void validate(String value) {
-        if (value.length() > LENGTH_LIMIT) {
-            throw new CustomException(ErrorCode.LENGTH_EXCEEDED, "아이템 제목은 " + LENGTH_LIMIT + "자를 넘을 수 없습니다.");
+        if (value != null && value.length() > LENGTH_LIMIT) {
+            throw new CustomException(ErrorCode.LENGTH_EXCEEDED, "아이템의 이미지 URL은 " + LENGTH_LIMIT + "자를 넘을 수 없습니다.");
         }
     }
 }
