+++ src/main/java/com/listywave/list/application/domain/item/ItemTitle.java
@@ -1,33 +1,35 @@
-package com.listywave.list.application.vo;
+package com.listywave.list.application.domain.item;
 
 import com.listywave.common.exception.CustomException;
 import com.listywave.common.exception.ErrorCode;
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
 import lombok.AccessLevel;
+import lombok.Builder;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
+@Builder
 @Embeddable
 @EqualsAndHashCode
 @NoArgsConstructor(access = AccessLevel.PROTECTED, force = true)
-public class Content {
+public class ItemTitle {
 
-    private static final int LENGTH_LIMIT = 500;
+    private static final int LENGTH_LIMIT = 100;
 
-    @Column(name = "content", nullable = false, length = LENGTH_LIMIT)
+    @Column(name = "title", nullable = false, length = LENGTH_LIMIT)
     private final String value;
 
-    public Content(String value) {
+    public ItemTitle(String value) {
         validate(value);
         this.value = value;
     }
 
     private void validate(String value) {
         if (value.length() > LENGTH_LIMIT) {
-            throw new CustomException(ErrorCode.LENGTH_EXCEEDED, "댓글은 " + LENGTH_LIMIT + "자를 넘을 수 없습니다.");
+            throw new CustomException(ErrorCode.LENGTH_EXCEEDED, "아이템 제목은 " + LENGTH_LIMIT + "자를 넘을 수 없습니다.");
         }
     }
 }
