+++ src/main/java/com/listywave/list/application/domain/list/Content.java
@@ -1,35 +1,33 @@
-package com.listywave.list.application.vo;
+package com.listywave.list.application.domain.list;
 
 import com.listywave.common.exception.CustomException;
 import com.listywave.common.exception.ErrorCode;
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
 import lombok.AccessLevel;
-import lombok.Builder;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
-@Builder
 @Embeddable
 @EqualsAndHashCode
 @NoArgsConstructor(access = AccessLevel.PROTECTED, force = true)
-public class ListDescription {
+public class Content {
 
-    private static final int LENGTH_LIMIT = 200;
+    private static final int LENGTH_LIMIT = 500;
 
-    @Column(name = "description", length = LENGTH_LIMIT)
+    @Column(name = "content", nullable = false, length = LENGTH_LIMIT)
     private final String value;
 
-    public ListDescription(String value) {
+    public Content(String value) {
         validate(value);
         this.value = value;
     }
 
     private void validate(String value) {
-        if (value != null && value.length() > LENGTH_LIMIT) {
-            throw new CustomException(ErrorCode.LENGTH_EXCEEDED, "리스트 설명은 " + LENGTH_LIMIT + "자를 넘을 수 없습니다.");
+        if (value.length() > LENGTH_LIMIT) {
+            throw new CustomException(ErrorCode.LENGTH_EXCEEDED, "댓글은 " + LENGTH_LIMIT + "자를 넘을 수 없습니다.");
         }
     }
 }
