+++ src/main/java/com/listywave/list/application/domain/list/ListDescription.java
@@ -1,4 +1,4 @@
-package com.listywave.list.application.vo;
+package com.listywave.list.application.domain.list;
 
 import com.listywave.common.exception.CustomException;
 import com.listywave.common.exception.ErrorCode;
@@ -15,21 +15,21 @@ import lombok.NoArgsConstructor;
 @Embeddable
 @EqualsAndHashCode
 @NoArgsConstructor(access = AccessLevel.PROTECTED, force = true)
-public class ItemLink {
+public class ListDescription {
 
-    private static final int LENGTH_LIMIT = 2048;
+    private static final int LENGTH_LIMIT = 200;
 
-    @Column(name = "link", length = LENGTH_LIMIT)
+    @Column(name = "description", length = LENGTH_LIMIT)
     private final String value;
 
-    public ItemLink(String value) {
+    public ListDescription(String value) {
         validate(value);
         this.value = value;
     }
 
     private void validate(String value) {
         if (value != null && value.length() > LENGTH_LIMIT) {
-            throw new CustomException(ErrorCode.LENGTH_EXCEEDED, "첨부 link 길이는 " + LENGTH_LIMIT + "자를 넘을 수 없습니다.");
+            throw new CustomException(ErrorCode.LENGTH_EXCEEDED, "리스트 설명은 " + LENGTH_LIMIT + "자를 넘을 수 없습니다.");
         }
     }
 }
