+++ src/main/java/com/listywave/list/application/dto/response/CommentFindResponse.java
@@ -2,8 +2,8 @@ package com.listywave.list.application.dto.response;
 
 import static java.util.Collections.emptyList;
 
-import com.listywave.list.application.domain.Comment;
-import com.listywave.list.application.domain.Reply;
+import com.listywave.list.application.domain.comment.Comment;
+import com.listywave.list.application.domain.reply.Reply;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Map;
