+++ src/main/java/com/listywave/list/application/dto/response/ListSearchResponse.java
@@ -1,7 +1,7 @@
 package com.listywave.list.application.dto.response;
 
-import com.listywave.list.application.domain.Item;
-import com.listywave.list.application.domain.ListEntity;
+import com.listywave.list.application.domain.item.Item;
+import com.listywave.list.application.domain.list.ListEntity;
 import java.time.LocalDateTime;
 import java.util.List;
 import lombok.Builder;
