+++ src/main/java/com/listywave/list/application/dto/response/ReplyCreateResponse.java
@@ -1,7 +1,7 @@
 package com.listywave.list.application.dto.response;
 
-import com.listywave.list.application.domain.Comment;
-import com.listywave.list.application.domain.Reply;
+import com.listywave.list.application.domain.comment.Comment;
+import com.listywave.list.application.domain.reply.Reply;
 import com.listywave.user.application.domain.User;
 import java.time.LocalDateTime;
 import lombok.Builder;
