+++ src/main/java/com/listywave/list/application/service/CommentService.java
@@ -6,9 +6,9 @@ import static java.util.stream.Collectors.toMap;
 
 import com.listywave.auth.application.domain.JwtManager;
 import com.listywave.common.exception.CustomException;
-import com.listywave.list.application.domain.Comment;
-import com.listywave.list.application.domain.ListEntity;
-import com.listywave.list.application.domain.Reply;
+import com.listywave.list.application.domain.comment.Comment;
+import com.listywave.list.application.domain.list.ListEntity;
+import com.listywave.list.application.domain.reply.Reply;
 import com.listywave.list.application.dto.response.CommentCreateResponse;
 import com.listywave.list.application.dto.response.CommentFindResponse;
 import com.listywave.list.repository.CommentRepository;
@@ -80,7 +80,7 @@ public class CommentService {
         if (!comment.canDeleteBy(user)) {
             throw new CustomException(INVALID_ACCESS, "댓글은 작성자만 지울 수 있습니다.");
         }
-        
+
         if (replyRepository.existsByComment(comment)) {
             comment.softDelete();
             return;
