+++ src/main/java/com/listywave/list/repository/CustomCommentRepositoryImpl.java
@@ -3,8 +3,8 @@ package com.listywave.list.repository;
 import static com.listywave.list.application.domain.QComment.comment;
 import static com.listywave.list.application.domain.QListEntity.listEntity;
 
-import com.listywave.list.application.domain.Comment;
-import com.listywave.list.application.domain.ListEntity;
+import com.listywave.list.application.domain.comment.Comment;
+import com.listywave.list.application.domain.list.ListEntity;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
