+++ src/main/java/com/listywave/list/repository/list/ListRepository.java
@@ -3,7 +3,7 @@ package com.listywave.list.repository.list;
 import static com.listywave.common.exception.ErrorCode.RESOURCE_NOT_FOUND;
 
 import com.listywave.common.exception.CustomException;
-import com.listywave.list.application.domain.ListEntity;
+import com.listywave.list.application.domain.list.ListEntity;
 import com.listywave.list.repository.list.custom.CustomListRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
