+++ src/main/java/com/listywave/list/repository/reply/CustomReplyRepositoryImpl.java
@@ -4,7 +4,7 @@ import static com.listywave.list.application.domain.QComment.comment;
 import static com.listywave.list.application.domain.QListEntity.listEntity;
 import static com.listywave.list.application.domain.QReply.reply;
 
-import com.listywave.list.application.domain.ListEntity;
+import com.listywave.list.application.domain.list.ListEntity;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
