+++ src/main/java/com/listywave/user/application/service/UserService.java
@@ -4,8 +4,8 @@ import com.listywave.auth.application.domain.JwtManager;
 import com.listywave.collaborator.application.domain.Collaborator;
 import com.listywave.collaborator.repository.CollaboratorRepository;
 import com.listywave.common.util.UserUtil;
-import com.listywave.list.application.domain.CategoryType;
-import com.listywave.list.application.domain.ListEntity;
+import com.listywave.list.application.domain.category.CategoryType;
+import com.listywave.list.application.domain.list.ListEntity;
 import com.listywave.user.application.domain.Follow;
 import com.listywave.user.application.domain.User;
 import com.listywave.user.application.dto.AllUserListsResponse;
