+++ src/main/java/com/listywave/user/repository/user/custom/impl/CustomUserRepositoryImpl.java
@@ -6,8 +6,8 @@ import static com.listywave.user.application.domain.QUser.user;
 
 import com.listywave.collaborator.application.domain.Collaborator;
 import com.listywave.collaborator.application.dto.CollaboratorResponse;
-import com.listywave.list.application.domain.CategoryType;
-import com.listywave.list.application.domain.ListEntity;
+import com.listywave.list.application.domain.category.CategoryType;
+import com.listywave.list.application.domain.list.ListEntity;
 import com.listywave.user.application.domain.User;
 import com.listywave.user.repository.user.custom.CustomUserRepository;
 import com.querydsl.core.types.Projections;
